Move autodiff_overloads from math to common
Scalartypes should live in common, and be grouped by scalartype.
