+++ drake/common/CMakeLists.txt
@@ -27,6 +27,7 @@ set(sources
 # List headers that should be installed with Drake so that they
 # are available elsewhere via #include "drake/common/xxx.h".
 set(installed_headers
+  autodiff_overloads.h
   constants.h
   drake_assert.h
   drake_deprecated.h
