+++ drake/common/autodiff_overloads.h
