+++ drake/math/CMakeLists.txt
@@ -20,7 +20,6 @@ set(sources
 set(installed_headers
   autodiff.h
   autodiff_gradient.h
-  autodiff_overloads.h
   axis_angle.h
   cross_product.h
   eigen_sparse_triplet.h
