+++ drake/math/expmap.h
@@ -7,10 +7,10 @@
 
 #include <Eigen/Dense>
 
+#include "drake/common/autodiff_overloads.h"
 #include "drake/common/drake_assert.h"
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
-#include "drake/math/autodiff_overloads.h"
 
 namespace drake {
 namespace math {
